keylogger/keylogger_project/requirements.txt
pynput
Pillow
requests
